Add ability to lookup single image. Only delete image resource in prod

+++ java-components/management-console/src/main/java/com/redhat/hacbs/management/watcher/JvmImageScanWatcher.java
@@ -179,7 +179,9 @@ public class JvmImageScanWatcher {
         }
         containerImage.analysisComplete = true;
         containerImage.persist();
-        client.resource(resource).delete();
+        if (LaunchMode.current() == LaunchMode.NORMAL) {
+            Log.infof("Marking image resource %s for deletion", resource.getSpec().getImage());
+            client.resource(resource).delete();
+        }
     }
-
 }
